feat: remove class-based events, add event dispatcher based on signals, rewrite input manager

+++ source/zyeware/utils/codes.d
@@ -385,7 +385,7 @@ enum MouseCode : ushort
     buttonMiddle = button2
 }
 
-alias GamepadIndex = ushort;
+alias GamepadIndex = size_t;
 
 /// Represents an abstract Xbox 360 gamepad button.
 enum GamepadButton
